# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Edgebook",
    "description": "College Football Paper-Betting Platform API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/ingestion/runs": {
      "get": {
        "tags": [
          "ingestion"
        ],
        "summary": "List Runs Endpoint",
        "description": "List ingestion run history newest-first.",
        "operationId": "list_runs_endpoint_ingestion_runs_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/ingestion/conflicts/{game_id}/resolve": {
      "post": {
        "tags": [
          "ingestion"
        ],
        "summary": "Resolve Conflict Endpoint",
        "description": "Resolve a score conflict for a specific game.",
        "operationId": "resolve_conflict_endpoint_ingestion_conflicts__game_id__resolve_post",
        "parameters": [
          {
            "name": "game_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Game Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConflictResolutionCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictResolutionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConflictResolutionCreate": {
        "properties": {
          "home_score": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Home Score"
          },
          "away_score": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Away Score"
          },
          "reason": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 1,
            "title": "Reason"
          },
          "resolved_by": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Resolved By"
          }
        },
        "type": "object",
        "required": [
          "home_score",
          "away_score",
          "reason",
          "resolved_by"
        ],
        "title": "ConflictResolutionCreate"
      },
      "ConflictResolutionResponse": {
        "properties": {
          "game_id": {
            "type": "string",
            "title": "Game Id"
          },
          "home_score": {
            "type": "integer",
            "title": "Home Score"
          },
          "away_score": {
            "type": "integer",
            "title": "Away Score"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "resolved_by": {
            "type": "string",
            "title": "Resolved By"
          },
          "resolved_at": {
            "type": "string",
            "title": "Resolved At"
          }
        },
        "type": "object",
        "required": [
          "game_id",
          "home_score",
          "away_score",
          "reason",
          "resolved_by",
          "resolved_at"
        ],
        "title": "ConflictResolutionResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RunPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RunResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "RunPage"
      },
      "RunResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "scope": {
            "type": "string",
            "title": "Scope"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "records_seen": {
            "type": "integer",
            "title": "Records Seen"
          },
          "records_created": {
            "type": "integer",
            "title": "Records Created"
          },
          "records_skipped": {
            "type": "integer",
            "title": "Records Skipped"
          },
          "conflict_count": {
            "type": "integer",
            "title": "Conflict Count"
          },
          "retry_count": {
            "type": "integer",
            "title": "Retry Count"
          },
          "requested_snapshot_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requested Snapshot At"
          },
          "provider_snapshot_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Snapshot At"
          },
          "quota_used": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quota Used"
          },
          "quota_remaining": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quota Remaining"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "provider",
          "scope",
          "status",
          "records_seen",
          "records_created",
          "records_skipped",
          "conflict_count",
          "retry_count",
          "requested_snapshot_at",
          "provider_snapshot_at",
          "quota_used",
          "quota_remaining",
          "error_message",
          "started_at",
          "finished_at"
        ],
        "title": "RunResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}